+++ packages/@toss/design-token/src/themes/theme.ts
@@ -1,7 +1,7 @@
 import { colors } from "./colors";
 import { radius } from "./radius";
 import { shadow } from "./shadow";
-import { zIndex } from "./zindex";
+import { zIndex } from "./zIndex";
 
 export const theme = {
   colors,
